Fix crash on Mac OS with fieldnames longer than 7 characters
On Mac OS, strcpy under the hood calls `__strcpy_chk` which attempts
to check if the destination is large enough to fit the input. Due to
how the `SHORTCUTLEAF` struct is defined in `JudySL.c`, `__strcpy_chk`
believes it can only fit `sizeof(WORDSIZE)`, while it is actually a
variable-sized field.

This change passes `_FORTIFY_SOURCE=0` to clang, which ensures
`__strcpy_chk` is *not* inserted.

Fixes https://github.com/traildb/traildb/issues/136